+++ src/component/countries/CountriesList.js
@@ -6,23 +6,21 @@ const CountriesList = () => {
   const stats = useSelector((state) => state.statsReducer);
 
   const list = stats ? stats.filter((item) => item.id !== 'selected' && item.id !== 'Total').map((item) => (
-		  <li key={item.id}>
-     <County 
-		 countryName={item.name}
-		 confimedCases={item.todayConfirmed}
-		 />
-		</li>
-	)) : [];
+    <li key={item.id}>
+      <County
+        countryName={item.name}
+        confimedCases={item.todayConfirmed}
+      />
+    </li>
+  )) : [];
 
-	
+  return (
 
-	return (
-		
-	 <ul className="countries_list">
-		{ list }
-	 </ul>
-	
-	)
-}
+    <ul className="countries_list">
+      { list }
+    </ul>
 
-export default CountriesList
+  );
+};
+
+export default CountriesList;
